moved useCountUp hook to it's own file
